# Optimize for-loop merging of cohorts.